Adds notifications to the backend
This is a change to how the backend handles subscriptions, now to
accomodate for notifications as well.
The current flow is:

A subscription is created for each user-subscription pair
Whenever a subscription is updated (which happens in two different ways:
discussion created, and discussion commented), a notification will be
generated
This notification, in turn, sends an email to the user
This means we'll now be notified by email not only on new comments, but
also when discussions are created.
The main reason for this is to allow for in-app notifications as well

Probably some people will complain about the spam this generates. A
temporary solution for that is for each of us to use gmail filters.
Other solutions can be thought about later