Makes discussion appear "READ" after being open
The status property needs to be added to the DiscussionSerializer.

The problem here is that we are making two requests to get all
information on a discussion.

* The first one is on discussions/index, loading a small ammount of
information for each discussions.

* The second one is for discussions/show. This requests ignores the
cached values and loads again all the information needed for a single discussion (except the status).

Once someone opens a discussions/show a request is made to mark that page
as read. After that request we need to refresh the information we have
on the discussion. For this we call the discussions/show again (not the
discussions/index).
For this reason we need it to return status.