Fixes notifications on new comment
The previous implementation relied on `touch: true` to automatically
save the discussion when a new comment was created. Aparently, this does
not trigger active record callbacks, so subscriptions where not being
updated.

Instead of adding yet another active_record callback, and starting to
get too much spaghetti there, I decided to move this to the controller
action. I also added a test for this