[Models] Improves subscription update system
Gets all the subscriptions in bulk and searches in memory, instead of a
query per user. Also calls the subscription set_initial_state callback
before save, instead of after, thus saving an access to the db per new
subscription created.

Moves the logic of creating subscriptions into it's own object. This
still does not deal with users unsubscribing to discussion, but it's a
step towards that.