[Serializers] Adds decorator for discussion with subscription
This a layer of logic to discussions, without impacting it directly.

The class method in the decorator has to be declared so it responds with
discussion and plays nicely with Rails.

I couldn't understand why I had to do the delegation to object in the
serializers, but it wouldn't work otherwise.