Removes compact serializer
Removing this serializer allows to reduce the number of requests and the
complexity. I know we can be facing an issue once this gets big, but I
believe we should then find a better solution. For now, I believe it's
better to have it this way.

One solution may be to just load a few discussions to the index and add
a load more button.