Replace material toolbar
This pr is a small first step to remove angular material. The directive I created is easier to use and extend.

Creates custom toolbar component and replaces everywhere.
Fixes issue when refreshing browser on /discussions/new page.
The reason the default is not having shadow is so we don't have overrides. We add shadows when we need them, instead of overriding to remove them.